added output

+++ xdecoder/architectures/xdecoder_model.py
@@ -524,7 +524,10 @@ class GeneralizedXdecoder(nn.Module):
         if keep_sem_bgd:
             mask_cls = F.softmax(mask_cls, dim=-1)
         else:
-            mask_cls = F.softmax(mask_cls, dim=-1)[..., :-1]
+            mask_cls = F.softmax(mask_cls, dim=-1, dtype=torch.float)[..., :-1]
+        # changed
+        mask_pred = mask_pred.type(torch.float)
+        # print(type(mask_pred))
         mask_pred = mask_pred.sigmoid()
         semseg = torch.einsum("qc,qhw->chw", mask_cls, mask_pred)
         return semseg
